Restricted Pages and User Registration

+++ base/views.py
@@ -1,8 +1,11 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -11,9 +14,12 @@ from .forms import RoomForm
 
 # Login Page
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -29,7 +35,7 @@ def loginPage(request):
         else:
             messages.error(request, 'Username or Password does not exists')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 
@@ -37,6 +43,26 @@ def loginPage(request):
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+
+# Register User
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occurred during registration')
+
+    context = {'form': form}
+    return render(request, 'base/login_register.html', context)
 
 
 # Home Page
@@ -66,6 +92,7 @@ def room(request, pk):
 
 
 # Create Room
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
 
@@ -80,9 +107,13 @@ def createRoom(request):
 
 
 # Update Room
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -96,8 +127,12 @@ def updateRoom(request, pk):
 
 
 # Delete Room
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
 
     if request.method == 'POST':
         room.delete()
